Refactor: Change to dataclass, Move files, Refactor test codes and Fix mypy (#14)
* Change to dataclass, Move files, More refactor with test codes and fix one mypy errors

* Update types
