+++ money/constants.py
@@ -1,0 +1,1559 @@
+from money.dataclasses.currency import Currency
+
+DEFAULT_CURRENCY = Currency(
+    code="XXX",
+    numeric="999",
+    decimals=0,
+    symbol="",
+    name="The codes assigned for transactions where no currency is involved",
+    countries=["ZZ07_No_Currency"],
+)
+
+# Definitions of ISO 4217 Currencies
+# Source: http://www.currency-iso.org/
+# Symbols: http://www.xe.com/symbols.php
+#
+# Note that the decimal code of N/A has been mapped to None
+CURRENCY_LIST: tuple[Currency, ...] = (
+    Currency(
+        code="AED",
+        numeric="784",
+        decimals=2,
+        symbol="",
+        name="UAE Dirham",
+        countries=["UNITED ARAB EMIRATES"],
+    ),
+    Currency(
+        code="AFN",
+        numeric="971",
+        decimals=2,
+        symbol="؋",
+        name="Afghani",
+        countries=["AFGHANISTAN"],
+    ),
+    Currency(
+        code="ALL",
+        numeric="008",
+        decimals=2,
+        symbol="Lek",
+        name="Lek",
+        countries=["ALBANIA"],
+    ),
+    Currency(
+        code="AMD",
+        numeric="051",
+        decimals=2,
+        symbol="",
+        name="Armenian Dram",
+        countries=["ARMENIA"],
+    ),
+    Currency(
+        code="ANG",
+        numeric="532",
+        decimals=2,
+        symbol="ƒ",
+        name="Netherlands Antillean Guilder",
+        countries=["CURA\xc7AO", "SINT MAARTEN (DUTCH PART)"],
+    ),
+    Currency(
+        code="AOA",
+        numeric="973",
+        decimals=2,
+        symbol="",
+        name="Kwanza",
+        countries=["ANGOLA"],
+    ),
+    Currency(
+        code="ARS",
+        numeric="032",
+        decimals=2,
+        symbol="$",
+        name="Argentine Peso",
+        countries=["ARGENTINA"],
+    ),
+    Currency(
+        code="AUD",
+        numeric="036",
+        decimals=2,
+        symbol="$",
+        name="Australian Dollar",
+        countries=[
+            "AUSTRALIA",
+            "CHRISTMAS ISLAND",
+            "COCOS (KEELING) ISLANDS",
+            "HEARD ISLAND AND McDONALD ISLANDS",
+            "KIRIBATI",
+            "NAURU",
+            "NORFOLK ISLAND",
+            "TUVALU",
+        ],
+    ),
+    Currency(
+        code="AWG",
+        numeric="533",
+        decimals=2,
+        symbol="ƒ",
+        name="Aruban Florin",
+        countries=["ARUBA"],
+    ),
+    Currency(
+        code="AZN",
+        numeric="944",
+        decimals=2,
+        symbol="ман",
+        name="Azerbaijanian Manat",
+        countries=["AZERBAIJAN"],
+    ),
+    Currency(
+        code="BAM",
+        numeric="977",
+        decimals=2,
+        symbol="KM",
+        name="Convertible Mark",
+        countries=["BOSNIA AND HERZEGOVINA"],
+    ),
+    Currency(
+        code="BBD",
+        numeric="052",
+        decimals=2,
+        symbol="$",
+        name="Barbados Dollar",
+        countries=["BARBADOS"],
+    ),
+    Currency(
+        code="BDT",
+        numeric="050",
+        decimals=2,
+        symbol="",
+        name="Taka",
+        countries=["BANGLADESH"],
+    ),
+    Currency(
+        code="BGN",
+        numeric="975",
+        decimals=2,
+        symbol="лв",
+        name="Bulgarian Lev",
+        countries=["BULGARIA"],
+    ),
+    Currency(
+        code="BHD",
+        numeric="048",
+        decimals=3,
+        symbol="",
+        name="Bahraini Dinar",
+        countries=["BAHRAIN"],
+    ),
+    Currency(
+        code="BIF",
+        numeric="108",
+        decimals=0,
+        symbol="",
+        name="Burundi Franc",
+        countries=["BURUNDI"],
+    ),
+    Currency(
+        code="BMD",
+        numeric="060",
+        decimals=2,
+        symbol="$",
+        name="Bermudian Dollar",
+        countries=["BERMUDA"],
+    ),
+    Currency(
+        code="BND",
+        numeric="096",
+        decimals=2,
+        symbol="$",
+        name="Brunei Dollar",
+        countries=["BRUNEI DARUSSALAM"],
+    ),
+    Currency(
+        code="BOB",
+        numeric="068",
+        decimals=2,
+        symbol="$b",
+        name="Boliviano",
+        countries=["BOLIVIA, PLURINATIONAL STATE OF"],
+    ),
+    Currency(
+        code="BOV",
+        numeric="984",
+        decimals=2,
+        symbol="",
+        name="Mvdol",
+        countries=["BOLIVIA, PLURINATIONAL STATE OF"],
+    ),
+    Currency(
+        code="BRL",
+        numeric="986",
+        decimals=2,
+        symbol="R$",
+        name="Brazilian Real",
+        countries=["BRAZIL"],
+    ),
+    Currency(
+        code="BSD",
+        numeric="044",
+        decimals=2,
+        symbol="$",
+        name="Bahamian Dollar",
+        countries=["BAHAMAS"],
+    ),
+    Currency(
+        code="BTN",
+        numeric="064",
+        decimals=2,
+        symbol="",
+        name="Ngultrum",
+        countries=["BHUTAN"],
+    ),
+    Currency(
+        code="BWP",
+        numeric="072",
+        decimals=2,
+        symbol="P",
+        name="Pula",
+        countries=["BOTSWANA"],
+    ),
+    Currency(
+        code="BYR",
+        numeric="974",
+        decimals=0,
+        symbol="p.",
+        name="Belarussian Ruble",
+        countries=["BELARUS"],
+    ),
+    Currency(
+        code="BZD",
+        numeric="084",
+        decimals=2,
+        symbol="BZ$",
+        name="Belize Dollar",
+        countries=["BELIZE"],
+    ),
+    Currency(
+        code="CAD",
+        numeric="124",
+        decimals=2,
+        symbol="$",
+        name="Canadian Dollar",
+        countries=["CANADA"],
+    ),
+    Currency(
+        code="CDF",
+        numeric="976",
+        decimals=2,
+        symbol="",
+        name="Congolese Franc",
+        countries=["CONGO, THE DEMOCRATIC REPUBLIC OF"],
+    ),
+    Currency(
+        code="CHE",
+        numeric="947",
+        decimals=2,
+        symbol="",
+        name="WIR Euro",
+        countries=["SWITZERLAND"],
+    ),
+    Currency(
+        code="CHF",
+        numeric="756",
+        decimals=2,
+        symbol="Fr.",
+        name="Swiss Franc",
+        countries=["LIECHTENSTEIN", "SWITZERLAND"],
+    ),
+    Currency(
+        code="CHW",
+        numeric="948",
+        decimals=2,
+        symbol="",
+        name="WIR Franc",
+        countries=["SWITZERLAND"],
+    ),
+    Currency(
+        code="CLF",
+        numeric="990",
+        decimals=0,
+        symbol="",
+        name="Unidades de fomento",
+        countries=["CHILE"],
+    ),
+    Currency(
+        code="CLP",
+        numeric="152",
+        decimals=0,
+        symbol="$",
+        name="Chilean Peso",
+        countries=["CHILE"],
+    ),
+    Currency(
+        code="CNY",
+        numeric="156",
+        decimals=2,
+        symbol="¥",
+        name="Yuan Renminbi",
+        countries=["CHINA"],
+    ),
+    Currency(
+        code="COP",
+        numeric="170",
+        decimals=2,
+        symbol="$",
+        name="Colombian Peso",
+        countries=["COLOMBIA"],
+    ),
+    Currency(
+        code="COU",
+        numeric="970",
+        decimals=2,
+        symbol="",
+        name="Unidad de Valor Real",
+        countries=["COLOMBIA"],
+    ),
+    Currency(
+        code="CRC",
+        numeric="188",
+        decimals=2,
+        symbol="₡",
+        name="Costa Rican Colon",
+        countries=["COSTA RICA"],
+    ),
+    Currency(
+        code="CUC",
+        numeric="931",
+        decimals=2,
+        symbol="",
+        name="Peso Convertible",
+        countries=["CUBA"],
+    ),
+    Currency(
+        code="CUP",
+        numeric="192",
+        decimals=2,
+        symbol="₱",
+        name="Cuban Peso",
+        countries=["CUBA"],
+    ),
+    Currency(
+        code="CVE",
+        numeric="132",
+        decimals=2,
+        symbol="",
+        name="Cape Verde Escudo",
+        countries=["CAPE VERDE"],
+    ),
+    Currency(
+        code="CZK",
+        numeric="203",
+        decimals=2,
+        symbol="Kč",
+        name="Czech Koruna",
+        countries=["CZECH REPUBLIC"],
+    ),
+    Currency(
+        code="DJF",
+        numeric="262",
+        decimals=0,
+        symbol="",
+        name="Djibouti Franc",
+        countries=["DJIBOUTI"],
+    ),
+    Currency(
+        code="DKK",
+        numeric="208",
+        decimals=2,
+        symbol="kr",
+        name="Danish Krone",
+        countries=["DENMARK", "FAROE ISLANDS", "GREENLAND"],
+    ),
+    Currency(
+        code="DOP",
+        numeric="214",
+        decimals=2,
+        symbol="RD$",
+        name="Dominican Peso",
+        countries=["DOMINICAN REPUBLIC"],
+    ),
+    Currency(
+        code="DZD",
+        numeric="012",
+        decimals=2,
+        symbol="",
+        name="Algerian Dinar",
+        countries=["ALGERIA"],
+    ),
+    Currency(
+        code="EGP",
+        numeric="818",
+        decimals=2,
+        symbol="£",
+        name="Egyptian Pound",
+        countries=["EGYPT"],
+    ),
+    Currency(
+        code="ERN",
+        numeric="232",
+        decimals=2,
+        symbol="",
+        name="Nakfa",
+        countries=["ERITREA"],
+    ),
+    Currency(
+        code="ETB",
+        numeric="230",
+        decimals=2,
+        symbol="",
+        name="Ethiopian Birr",
+        countries=["ETHIOPIA"],
+    ),
+    Currency(
+        code="EUR",
+        numeric="978",
+        decimals=2,
+        symbol="€",
+        name="Euro",
+        countries=[
+            "\xc5LAND ISLANDS",
+            "ANDORRA",
+            "AUSTRIA",
+            "BELGIUM",
+            "CYPRUS",
+            "ESTONIA",
+            "EUROPEAN UNION ",
+            "FINLAND",
+            "FRANCE",
+            "FRENCH GUIANA",
+            "FRENCH SOUTHERN TERRITORIES",
+            "GERMANY",
+            "GREECE",
+            "GUADELOUPE",
+            "HOLY SEE (VATICAN CITY STATE)",
+            "IRELAND",
+            "ITALY",
+            "LUXEMBOURG",
+            "MALTA",
+            "MARTINIQUE",
+            "MAYOTTE",
+            "MONACO",
+            "MONTENEGRO",
+            "NETHERLANDS",
+            "PORTUGAL",
+            "R\xc9UNION",
+            "SAINT BARTH\xc9LEMY",
+            "SAINT MARTIN (FRENCH PART)",
+            "SAINT PIERRE AND MIQUELON",
+            "SAN MARINO",
+            "SLOVAKIA",
+            "SLOVENIA",
+            "SPAIN",
+            "Vatican City State (HOLY SEE)",
+        ],
+    ),
+    Currency(
+        code="FJD",
+        numeric="242",
+        decimals=2,
+        symbol="$",
+        name="Fiji Dollar",
+        countries=["FIJI"],
+    ),
+    Currency(
+        code="FKP",
+        numeric="238",
+        decimals=2,
+        symbol="£",
+        name="Falkland Islands Pound",
+        countries=["FALKLAND ISLANDS (MALVINAS)"],
+    ),
+    Currency(
+        code="GBP",
+        numeric="826",
+        decimals=2,
+        symbol="£",
+        name="Pound Sterling",
+        countries=["GUERNSEY", "ISLE OF MAN", "JERSEY", "UNITED KINGDOM"],
+    ),
+    Currency(
+        code="GEL",
+        numeric="981",
+        decimals=2,
+        symbol="",
+        name="Lari",
+        countries=["GEORGIA"],
+    ),
+    Currency(
+        code="GHS",
+        numeric="936",
+        decimals=2,
+        symbol="",
+        name="Ghana Cedi",
+        countries=["GHANA"],
+    ),
+    Currency(
+        code="GIP",
+        numeric="292",
+        decimals=2,
+        symbol="£",
+        name="Gibraltar Pound",
+        countries=["GIBRALTAR"],
+    ),
+    Currency(
+        code="GMD",
+        numeric="270",
+        decimals=2,
+        symbol="",
+        name="Dalasi",
+        countries=["GAMBIA"],
+    ),
+    Currency(
+        code="GNF",
+        numeric="324",
+        decimals=0,
+        symbol="",
+        name="Guinea Franc",
+        countries=["GUINEA"],
+    ),
+    Currency(
+        code="GTQ",
+        numeric="320",
+        decimals=2,
+        symbol="Q",
+        name="Quetzal",
+        countries=["GUATEMALA"],
+    ),
+    Currency(
+        code="GYD",
+        numeric="328",
+        decimals=2,
+        symbol="$",
+        name="Guyana Dollar",
+        countries=["GUYANA"],
+    ),
+    Currency(
+        code="HKD",
+        numeric="344",
+        decimals=2,
+        symbol="HK$",
+        name="Hong Kong Dollar",
+        countries=["HONG KONG"],
+    ),
+    Currency(
+        code="HNL",
+        numeric="340",
+        decimals=2,
+        symbol="L",
+        name="Lempira",
+        countries=["HONDURAS"],
+    ),
+    Currency(
+        code="HRK",
+        numeric="191",
+        decimals=2,
+        symbol="kn",
+        name="Croatian Kuna",
+        countries=["CROATIA"],
+    ),
+    Currency(
+        code="HTG",
+        numeric="332",
+        decimals=2,
+        symbol="",
+        name="Gourde",
+        countries=["HAITI"],
+    ),
+    Currency(
+        code="HUF",
+        numeric="348",
+        decimals=2,
+        symbol="Ft",
+        name="Forint",
+        countries=["HUNGARY"],
+    ),
+    Currency(
+        code="IDR",
+        numeric="360",
+        decimals=2,
+        symbol="Rp",
+        name="Rupiah",
+        countries=["INDONESIA"],
+    ),
+    Currency(
+        code="ILS",
+        numeric="376",
+        decimals=2,
+        symbol="₪",
+        name="New Israeli Sheqel",
+        countries=["ISRAEL"],
+    ),
+    Currency(
+        code="INR",
+        numeric="356",
+        decimals=2,
+        symbol="",
+        name="Indian Rupee",
+        countries=["BHUTAN", "INDIA"],
+    ),
+    Currency(
+        code="IQD",
+        numeric="368",
+        decimals=3,
+        symbol="",
+        name="Iraqi Dinar",
+        countries=["IRAQ"],
+    ),
+    Currency(
+        code="IRR",
+        numeric="364",
+        decimals=2,
+        symbol="﷼",
+        name="Iranian Rial",
+        countries=["IRAN, ISLAMIC REPUBLIC OF"],
+    ),
+    Currency(
+        code="ISK",
+        numeric="352",
+        decimals=0,
+        symbol="kr",
+        name="Iceland Krona",
+        countries=["ICELAND"],
+    ),
+    Currency(
+        code="JMD",
+        numeric="388",
+        decimals=2,
+        symbol="J$",
+        name="Jamaican Dollar",
+        countries=["JAMAICA"],
+    ),
+    Currency(
+        code="JOD",
+        numeric="400",
+        decimals=3,
+        symbol="",
+        name="Jordanian Dinar",
+        countries=["JORDAN"],
+    ),
+    Currency(
+        code="JPY",
+        numeric="392",
+        decimals=0,
+        symbol="¥",
+        name="Yen",
+        countries=["JAPAN"],
+    ),
+    Currency(
+        code="KES",
+        numeric="404",
+        decimals=2,
+        symbol="",
+        name="Kenyan Shilling",
+        countries=["KENYA"],
+    ),
+    Currency(
+        code="KGS",
+        numeric="417",
+        decimals=2,
+        symbol="лв",
+        name="Som",
+        countries=["KYRGYZSTAN"],
+    ),
+    Currency(
+        code="KHR",
+        numeric="116",
+        decimals=2,
+        symbol="៛",
+        name="Riel",
+        countries=["CAMBODIA"],
+    ),
+    Currency(
+        code="KMF",
+        numeric="174",
+        decimals=0,
+        symbol="",
+        name="Comoro Franc",
+        countries=["COMOROS"],
+    ),
+    Currency(
+        code="KPW",
+        numeric="408",
+        decimals=2,
+        symbol="₩",
+        name="North Korean Won",
+        countries=["KOREA, DEMOCRATIC PEOPLE\u2019S REPUBLIC OF"],
+    ),
+    Currency(
+        code="KRW",
+        numeric="410",
+        decimals=0,
+        symbol="₩",
+        name="Won",
+        countries=["KOREA, REPUBLIC OF"],
+    ),
+    Currency(
+        code="KWD",
+        numeric="414",
+        decimals=3,
+        symbol="",
+        name="Kuwaiti Dinar",
+        countries=["KUWAIT"],
+    ),
+    Currency(
+        code="KYD",
+        numeric="136",
+        decimals=2,
+        symbol="$",
+        name="Cayman Islands Dollar",
+        countries=["CAYMAN ISLANDS"],
+    ),
+    Currency(
+        code="KZT",
+        numeric="398",
+        decimals=2,
+        symbol="лв",
+        name="Tenge",
+        countries=["KAZAKHSTAN"],
+    ),
+    Currency(
+        code="LAK",
+        numeric="418",
+        decimals=2,
+        symbol="₭",
+        name="Kip",
+        countries=["LAO PEOPLE\u2019S DEMOCRATIC REPUBLIC"],
+    ),
+    Currency(
+        code="LBP",
+        numeric="422",
+        decimals=2,
+        symbol="£",
+        name="Lebanese Pound",
+        countries=["LEBANON"],
+    ),
+    Currency(
+        code="LKR",
+        numeric="144",
+        decimals=2,
+        symbol="₨",
+        name="Sri Lanka Rupee",
+        countries=["SRI LANKA"],
+    ),
+    Currency(
+        code="LRD",
+        numeric="430",
+        decimals=2,
+        symbol="$",
+        name="Liberian Dollar",
+        countries=["LIBERIA"],
+    ),
+    Currency(
+        code="LSL",
+        numeric="426",
+        decimals=2,
+        symbol="",
+        name="Loti",
+        countries=["LESOTHO"],
+    ),
+    Currency(
+        code="LTL",
+        numeric="440",
+        decimals=2,
+        symbol="Lt",
+        name="Lithuanian Litas",
+        countries=["LITHUANIA"],
+    ),
+    Currency(
+        code="LVL",
+        numeric="428",
+        decimals=2,
+        symbol="Ls",
+        name="Latvian Lats",
+        countries=["LATVIA"],
+    ),
+    Currency(
+        code="LYD",
+        numeric="434",
+        decimals=3,
+        symbol="",
+        name="Libyan Dinar",
+        countries=["LIBYA"],
+    ),
+    Currency(
+        code="MAD",
+        numeric="504",
+        decimals=2,
+        symbol="",
+        name="Moroccan Dirham",
+        countries=["MOROCCO", "WESTERN SAHARA"],
+    ),
+    Currency(
+        code="MDL",
+        numeric="498",
+        decimals=2,
+        symbol="",
+        name="Moldovan Leu",
+        countries=["MOLDOVA, REPUBLIC OF"],
+    ),
+    Currency(
+        code="MGA",
+        numeric="969",
+        decimals=2,
+        symbol="",
+        name="Malagasy Ariary",
+        countries=["MADAGASCAR"],
+    ),
+    Currency(
+        code="MKD",
+        numeric="807",
+        decimals=2,
+        symbol="ден",
+        name="Denar",
+        countries=["MACEDONIA, THE FORMER YUGOSLAV REPUBLIC OF"],
+    ),
+    Currency(
+        code="MMK",
+        numeric="104",
+        decimals=2,
+        symbol="K",
+        name="Kyat",
+        countries=["MYANMAR"],
+    ),
+    Currency(
+        code="MNT",
+        numeric="496",
+        decimals=2,
+        symbol="₮",
+        name="Tugrik",
+        countries=["MONGOLIA"],
+    ),
+    Currency(
+        code="MOP",
+        numeric="446",
+        decimals=2,
+        symbol="",
+        name="Pataca",
+        countries=["MACAO"],
+    ),
+    Currency(
+        code="MRO",
+        numeric="478",
+        decimals=2,
+        symbol="",
+        name="Ouguiya",
+        countries=["MAURITANIA"],
+    ),
+    Currency(
+        code="MUR",
+        numeric="480",
+        decimals=2,
+        symbol="₨",
+        name="Mauritius Rupee",
+        countries=["MAURITIUS"],
+    ),
+    Currency(
+        code="MVR",
+        numeric="462",
+        decimals=2,
+        symbol="",
+        name="Rufiyaa",
+        countries=["MALDIVES"],
+    ),
+    Currency(
+        code="MWK",
+        numeric="454",
+        decimals=2,
+        symbol="",
+        name="Kwacha",
+        countries=["MALAWI"],
+    ),
+    Currency(
+        code="MXN",
+        numeric="484",
+        decimals=2,
+        symbol="$",
+        name="Mexican Peso",
+        countries=["MEXICO"],
+    ),
+    Currency(
+        code="MXV",
+        numeric="979",
+        decimals=2,
+        symbol="",
+        name="Mexican Unidad de Inversion (UDI)",
+        countries=["MEXICO"],
+    ),
+    Currency(
+        code="MYR",
+        numeric="458",
+        decimals=2,
+        symbol="RM",
+        name="Malaysian Ringgit",
+        countries=["MALAYSIA"],
+    ),
+    Currency(
+        code="MZN",
+        numeric="943",
+        decimals=2,
+        symbol="MT",
+        name="Mozambique Metical",
+        countries=["MOZAMBIQUE"],
+    ),
+    Currency(
+        code="NAD",
+        numeric="516",
+        decimals=2,
+        symbol="$",
+        name="Namibia Dollar",
+        countries=["NAMIBIA"],
+    ),
+    Currency(
+        code="NGN",
+        numeric="566",
+        decimals=2,
+        symbol="₦",
+        name="Naira",
+        countries=["NIGERIA"],
+    ),
+    Currency(
+        code="NIO",
+        numeric="558",
+        decimals=2,
+        symbol="C$",
+        name="Cordoba Oro",
+        countries=["NICARAGUA"],
+    ),
+    Currency(
+        code="NOK",
+        numeric="578",
+        decimals=2,
+        symbol="kr",
+        name="Norwegian Krone",
+        countries=["BOUVET ISLAND", "NORWAY", "SVALBARD AND JAN MAYEN"],
+    ),
+    Currency(
+        code="NPR",
+        numeric="524",
+        decimals=2,
+        symbol="₨",
+        name="Nepalese Rupee",
+        countries=["NEPAL"],
+    ),
+    Currency(
+        code="NZD",
+        numeric="554",
+        decimals=2,
+        symbol="$",
+        name="New Zealand Dollar",
+        countries=["COOK ISLANDS", "NEW ZEALAND", "NIUE", "PITCAIRN", "TOKELAU"],
+    ),
+    Currency(
+        code="OMR",
+        numeric="512",
+        decimals=3,
+        symbol="﷼",
+        name="Rial Omani",
+        countries=["OMAN"],
+    ),
+    Currency(
+        code="PAB",
+        numeric="590",
+        decimals=2,
+        symbol="B/.",
+        name="Balboa",
+        countries=["PANAMA"],
+    ),
+    Currency(
+        code="PEN",
+        numeric="604",
+        decimals=2,
+        symbol="S/.",
+        name="Nuevo Sol",
+        countries=["PERU"],
+    ),
+    Currency(
+        code="PGK",
+        numeric="598",
+        decimals=2,
+        symbol="",
+        name="Kina",
+        countries=["PAPUA NEW GUINEA"],
+    ),
+    Currency(
+        code="PHP",
+        numeric="608",
+        decimals=2,
+        symbol="₱",
+        name="Philippine Peso",
+        countries=["PHILIPPINES"],
+    ),
+    Currency(
+        code="PKR",
+        numeric="586",
+        decimals=2,
+        symbol="₨",
+        name="Pakistan Rupee",
+        countries=["PAKISTAN"],
+    ),
+    Currency(
+        code="PLN",
+        numeric="985",
+        decimals=2,
+        symbol="zł",
+        name="Zloty",
+        countries=["POLAND"],
+    ),
+    Currency(
+        code="PYG",
+        numeric="600",
+        decimals=0,
+        symbol="Gs",
+        name="Guarani",
+        countries=["PARAGUAY"],
+    ),
+    Currency(
+        code="QAR",
+        numeric="634",
+        decimals=2,
+        symbol="﷼",
+        name="Qatari Rial",
+        countries=["QATAR"],
+    ),
+    Currency(
+        code="RON",
+        numeric="946",
+        decimals=2,
+        symbol="lei",
+        name="New Romanian Leu",
+        countries=["ROMANIA"],
+    ),
+    Currency(
+        code="RSD",
+        numeric="941",
+        decimals=2,
+        symbol="Дин.",
+        name="Serbian Dinar",
+        countries=["SERBIA "],
+    ),
+    Currency(
+        code="RUB",
+        numeric="643",
+        decimals=2,
+        symbol="руб",
+        name="Russian Ruble",
+        countries=["RUSSIAN FEDERATION"],
+    ),
+    Currency(
+        code="RWF",
+        numeric="646",
+        decimals=0,
+        symbol="",
+        name="Rwanda Franc",
+        countries=["RWANDA"],
+    ),
+    Currency(
+        code="SAR",
+        numeric="682",
+        decimals=2,
+        symbol="﷼",
+        name="Saudi Riyal",
+        countries=["SAUDI ARABIA"],
+    ),
+    Currency(
+        code="SBD",
+        numeric="090",
+        decimals=2,
+        symbol="$",
+        name="Solomon Islands Dollar",
+        countries=["SOLOMON ISLANDS"],
+    ),
+    Currency(
+        code="SCR",
+        numeric="690",
+        decimals=2,
+        symbol="₨",
+        name="Seychelles Rupee",
+        countries=["SEYCHELLES"],
+    ),
+    Currency(
+        code="SDG",
+        numeric="938",
+        decimals=2,
+        symbol="",
+        name="Sudanese Pound",
+        countries=["SUDAN"],
+    ),
+    Currency(
+        code="SEK",
+        numeric="752",
+        decimals=2,
+        symbol="kr",
+        name="Swedish Krona",
+        countries=["SWEDEN"],
+    ),
+    Currency(
+        code="SGD",
+        numeric="702",
+        decimals=2,
+        symbol="$",
+        name="Singapore Dollar",
+        countries=["SINGAPORE"],
+    ),
+    Currency(
+        code="SHP",
+        numeric="654",
+        decimals=2,
+        symbol="£",
+        name="Saint Helena Pound",
+        countries=["SAINT HELENA, ASCENSION AND TRISTAN DA CUNHA"],
+    ),
+    Currency(
+        code="SLL",
+        numeric="694",
+        decimals=2,
+        symbol="",
+        name="Leone",
+        countries=["SIERRA LEONE"],
+    ),
+    Currency(
+        code="SOS",
+        numeric="706",
+        decimals=2,
+        symbol="S",
+        name="Somali Shilling",
+        countries=["SOMALIA"],
+    ),
+    Currency(
+        code="SRD",
+        numeric="968",
+        decimals=2,
+        symbol="$",
+        name="Surinam Dollar",
+        countries=["SURINAME"],
+    ),
+    Currency(
+        code="SSP",
+        numeric="728",
+        decimals=2,
+        symbol="",
+        name="South Sudanese Pound",
+        countries=["SOUTH SUDAN"],
+    ),
+    Currency(
+        code="STD",
+        numeric="678",
+        decimals=2,
+        symbol="",
+        name="Dobra",
+        countries=["SAO TOME AND PRINCIPE"],
+    ),
+    Currency(
+        code="SVC",
+        numeric="222",
+        decimals=2,
+        symbol="$",
+        name="El Salvador Colon",
+        countries=["EL SALVADOR"],
+    ),
+    Currency(
+        code="SYP",
+        numeric="760",
+        decimals=2,
+        symbol="£",
+        name="Syrian Pound",
+        countries=["SYRIAN ARAB REPUBLIC"],
+    ),
+    Currency(
+        code="SZL",
+        numeric="748",
+        decimals=2,
+        symbol="",
+        name="Lilangeni",
+        countries=["SWAZILAND"],
+    ),
+    Currency(
+        code="THB",
+        numeric="764",
+        decimals=2,
+        symbol="฿",
+        name="Baht",
+        countries=["THAILAND"],
+    ),
+    Currency(
+        code="TJS",
+        numeric="972",
+        decimals=2,
+        symbol="",
+        name="Somoni",
+        countries=["TAJIKISTAN"],
+    ),
+    Currency(
+        code="TMT",
+        numeric="934",
+        decimals=2,
+        symbol="",
+        name="Turkmenistan New Manat",
+        countries=["TURKMENISTAN"],
+    ),
+    Currency(
+        code="TND",
+        numeric="788",
+        decimals=3,
+        symbol="",
+        name="Tunisian Dinar",
+        countries=["TUNISIA"],
+    ),
+    Currency(
+        code="TOP",
+        numeric="776",
+        decimals=2,
+        symbol="",
+        name="Pa’anga",
+        countries=["TONGA"],
+    ),
+    Currency(
+        code="TRY",
+        numeric="949",
+        decimals=2,
+        symbol="TL",
+        name="Turkish Lira",
+        countries=["TURKEY"],
+    ),
+    Currency(
+        code="TTD",
+        numeric="780",
+        decimals=2,
+        symbol="TT$",
+        name="Trinidad and Tobago Dollar",
+        countries=["TRINIDAD AND TOBAGO"],
+    ),
+    Currency(
+        code="TWD",
+        numeric="901",
+        decimals=2,
+        symbol="NT$",
+        name="New Taiwan Dollar",
+        countries=["TAIWAN, PROVINCE OF CHINA"],
+    ),
+    Currency(
+        code="TZS",
+        numeric="834",
+        decimals=2,
+        symbol="",
+        name="Tanzanian Shilling",
+        countries=["TANZANIA, UNITED REPUBLIC OF"],
+    ),
+    Currency(
+        code="UAH",
+        numeric="980",
+        decimals=2,
+        symbol="₴",
+        name="Hryvnia",
+        countries=["UKRAINE"],
+    ),
+    Currency(
+        code="UGX",
+        numeric="800",
+        decimals=2,
+        symbol="",
+        name="Uganda Shilling",
+        countries=["UGANDA"],
+    ),
+    Currency(
+        code="USD",
+        numeric="840",
+        decimals=2,
+        symbol="$",
+        name="US Dollar",
+        countries=[
+            "AMERICAN SAMOA",
+            "BONAIRE, SINT EUSTATIUS AND SABA",
+            "BRITISH INDIAN OCEAN TERRITORY",
+            "ECUADOR",
+            "EL SALVADOR",
+            "GUAM",
+            "HAITI",
+            "MARSHALL ISLANDS",
+            "MICRONESIA, FEDERATED STATES OF",
+            "NORTHERN MARIANA ISLANDS",
+            "PALAU",
+            "PANAMA",
+            "PUERTO RICO",
+            "TIMOR-LESTE",
+            "TURKS AND CAICOS ISLANDS",
+            "UNITED STATES",
+            "UNITED STATES MINOR OUTLYING ISLANDS",
+            "VIRGIN ISLANDS (BRITISH)",
+            "VIRGIN ISLANDS (US)",
+        ],
+    ),
+    Currency(
+        code="USN",
+        numeric="997",
+        decimals=2,
+        symbol="$",
+        name="US Dollar (Next day)",
+        countries=["UNITED STATES"],
+    ),
+    Currency(
+        code="USS",
+        numeric="998",
+        decimals=2,
+        symbol="$",
+        name="US Dollar (Same day)",
+        countries=["UNITED STATES"],
+    ),
+    Currency(
+        code="UYI",
+        numeric="940",
+        decimals=0,
+        symbol="",
+        name="Uruguay Peso en Unidades Indexadas (URUIURUI)",
+        countries=["URUGUAY"],
+    ),
+    Currency(
+        code="UYU",
+        numeric="858",
+        decimals=2,
+        symbol="$U",
+        name="Peso Uruguayo",
+        countries=["URUGUAY"],
+    ),
+    Currency(
+        code="UZS",
+        numeric="860",
+        decimals=2,
+        symbol="лв",
+        name="Uzbekistan Sum",
+        countries=["UZBEKISTAN"],
+    ),
+    Currency(
+        code="VEF",
+        numeric="937",
+        decimals=2,
+        symbol="Bs",
+        name="Bolivar Fuerte",
+        countries=["VENEZUELA, BOLIVARIAN REPUBLIC OF"],
+    ),
+    Currency(
+        code="VND",
+        numeric="704",
+        decimals=0,
+        symbol="₫",
+        name="Dong",
+        countries=["VIET NAM"],
+    ),
+    Currency(
+        code="VUV",
+        numeric="548",
+        decimals=0,
+        symbol="",
+        name="Vatu",
+        countries=["VANUATU"],
+    ),
+    Currency(
+        code="WST",
+        numeric="882",
+        decimals=2,
+        symbol="",
+        name="Tala",
+        countries=["SAMOA"],
+    ),
+    Currency(
+        code="XAF",
+        numeric="950",
+        decimals=0,
+        symbol="",
+        name="CFA Franc BEAC",
+        countries=[
+            "CAMEROON",
+            "CENTRAL AFRICAN REPUBLIC",
+            "CHAD",
+            "CONGO",
+            "EQUATORIAL GUINEA",
+            "GABON",
+        ],
+    ),
+    Currency(
+        code="XAG",
+        numeric="961",
+        decimals=0,
+        symbol="",
+        name="Silver",
+        countries=["ZZ11_Silver"],
+    ),
+    Currency(
+        code="XAU",
+        numeric="959",
+        decimals=0,
+        symbol="",
+        name="Gold",
+        countries=["ZZ08_Gold"],
+    ),
+    Currency(
+        code="XBA",
+        numeric="955",
+        decimals=0,
+        symbol="",
+        name="Bond Markets Unit European Composite Unit (EURCO)",
+        countries=["ZZ01_Bond Markets Unit European_EURCO"],
+    ),
+    Currency(
+        code="XBB",
+        numeric="956",
+        decimals=0,
+        symbol="",
+        name="Bond Markets Unit European Monetary Unit (E.M.U.-6)",
+        countries=["ZZ02_Bond Markets Unit European_EMU-6"],
+    ),
+    Currency(
+        code="XBC",
+        numeric="957",
+        decimals=0,
+        symbol="",
+        name="Bond Markets Unit European Unit of Account 9 (E.U.A.-9)",
+        countries=["ZZ03_Bond Markets Unit European_EUA-9"],
+    ),
+    Currency(
+        code="XBD",
+        numeric="958",
+        decimals=0,
+        symbol="",
+        name="Bond Markets Unit European Unit of Account 17 (E.U.A.-17)",
+        countries=["ZZ04_Bond Markets Unit European_EUA-17"],
+    ),
+    Currency(
+        code="XCD",
+        numeric="951",
+        decimals=2,
+        symbol="$",
+        name="East Caribbean Dollar",
+        countries=[
+            "ANGUILLA",
+            "ANTIGUA AND BARBUDA",
+            "DOMINICA",
+            "GRENADA",
+            "MONTSERRAT",
+            "SAINT KITTS AND NEVIS",
+            "SAINT LUCIA",
+            "SAINT VINCENT AND THE GRENADINES",
+        ],
+    ),
+    Currency(
+        code="XDR",
+        numeric="960",
+        decimals=0,
+        symbol="",
+        name="SDR (Special Drawing Right)",
+        countries=["INTERNATIONAL MONETARY FUND (IMF)\xa0"],
+    ),
+    Currency(
+        code="XFU",
+        numeric="Nil",
+        decimals=0,
+        symbol="",
+        name="UIC-Franc",
+        countries=["ZZ05_UIC-Franc"],
+    ),
+    Currency(
+        code="XOF",
+        numeric="952",
+        decimals=0,
+        symbol="",
+        name="CFA Franc BCEAO",
+        countries=[
+            "BENIN",
+            "BURKINA FASO",
+            "C\xd4TE D'IVOIRE",
+            "GUINEA-BISSAU",
+            "MALI",
+            "NIGER",
+            "SENEGAL",
+            "TOGO",
+        ],
+    ),
+    Currency(
+        code="XPD",
+        numeric="964",
+        decimals=0,
+        symbol="",
+        name="Palladium",
+        countries=["ZZ09_Palladium"],
+    ),
+    Currency(
+        code="XPF",
+        numeric="953",
+        decimals=0,
+        symbol="",
+        name="CFP Franc",
+        countries=["FRENCH POLYNESIA", "NEW CALEDONIA", "WALLIS AND FUTUNA"],
+    ),
+    Currency(
+        code="XPT",
+        numeric="962",
+        decimals=0,
+        symbol="",
+        name="Platinum",
+        countries=["ZZ10_Platinum"],
+    ),
+    Currency(
+        code="XSU",
+        numeric="994",
+        decimals=0,
+        symbol="",
+        name="Sucre",
+        countries=['SISTEMA UNITARIO DE COMPENSACION REGIONAL DE PAGOS "SUCRE" '],
+    ),
+    Currency(
+        code="XTS",
+        numeric="963",
+        decimals=0,
+        symbol="",
+        name="Codes specifically reserved for testing purposes",
+        countries=["ZZ06_Testing_Code"],
+    ),
+    Currency(
+        code="XUA",
+        numeric="965",
+        decimals=0,
+        symbol="",
+        name="ADB Unit of Account",
+        countries=["MEMBER COUNTRIES OF THE AFRICAN DEVELOPMENT BANK GROUP"],
+    ),
+    DEFAULT_CURRENCY,
+    Currency(
+        code="YER",
+        numeric="886",
+        decimals=2,
+        symbol="﷼",
+        name="Yemeni Rial",
+        countries=["YEMEN"],
+    ),
+    Currency(
+        code="ZAR",
+        numeric="710",
+        decimals=2,
+        symbol="R",
+        name="Rand",
+        countries=["LESOTHO", "NAMIBIA", "SOUTH AFRICA"],
+    ),
+    Currency(
+        code="ZMK",
+        numeric="894",
+        decimals=2,
+        symbol="",
+        name="Zambian Kwacha",
+        countries=["ZAMBIA"],
+    ),
+    Currency(
+        code="ZWL",
+        numeric="932",
+        decimals=2,
+        symbol="",
+        name="Zimbabwe Dollar",
+        countries=["ZIMBABWE"],
+    ),
+)
+
+CURRENCY: dict[str, Currency] = {currency.code: currency for currency in CURRENCY_LIST}
+assert len(CURRENCY) == len(CURRENCY_LIST)
