+++ money/contrib/django/forms/fields.py
@@ -1,6 +1,7 @@
 from django import forms
 
-from money.money import CURRENCY, Money
+from money.constants import CURRENCY
+from money.dataclasses.money import Money
 
 from .widgets import CurrencySelectWidget
 
@@ -30,7 +31,7 @@ class MoneyField(forms.MultiValueField):
             ),
             forms.ChoiceField(choices=choices),
         )
-        super(MoneyField, self).__init__(fields, *args, **kwargs)
+        super().__init__(fields, *args, **kwargs)
 
     def compress(self, data_list):
         """
