+++ money/tests/test_django_defaults.py
@@ -1,6 +1,6 @@
 import pytest
 
-from money.money import Money
+from money.dataclasses.money import Money
 from money.tests.models import ALL_PARAMETRIZED_MODELS, ParametrizedModelType
